\name{return-codes}
\alias{return_codes}
\title{RApache handler return codes}
\description{
	RApache handlers must return one of the following return codes.
}
\details{ 
	The return codes are broken up into two sets. The first group is
	related to how the handler handles the request, and the second
	group are the valid HTTP response codes.
}
\section{First Group}{
\item{DONE}{The handler has \emph{completely} fulfilled the request, meaning no other request stage will alter it in any way. }
\item{DECLINED}{the handler has declined doing anything with the request}
\item{OK}{The handler has done its part; the request is sent to the other request stages. \emph{IN GENERAL, HANDLERS SHOULD RETURN THIS}}
}
\section{Second Group}{
	See RFC2616 for details on these.\cr\cr
HTTP\_CONTINUE \cr
HTTP\_SWITCHING\_PROTOCOLS \cr
HTTP\_PROCESSING \cr
HTTP\_OK \cr
HTTP\_CREATED \cr
HTTP\_ACCEPTED \cr
HTTP\_NON\_AUTHORITATIVE \cr
HTTP\_NO\_CONTENT \cr
HTTP\_RESET\_CONTENT \cr
HTTP\_PARTIAL\_CONTENT \cr
HTTP\_MULTI\_STATUS \cr
HTTP\_MULTIPLE\_CHOICES \cr
HTTP\_MOVED\_PERMANENTLY \cr
HTTP\_MOVED\_TEMPORARILY \cr
HTTP\_SEE\_OTHER \cr
HTTP\_NOT\_MODIFIED \cr
HTTP\_USE\_PROXY \cr
HTTP\_TEMPORARY\_REDIRECT \cr
HTTP\_BAD\_REQUEST \cr
HTTP\_UNAUTHORIZED \cr
HTTP\_PAYMENT\_REQUIRED \cr
HTTP\_FORBIDDEN \cr
HTTP\_NOT\_FOUND \cr
HTTP\_METHOD\_NOT\_ALLOWED \cr
HTTP\_NOT\_ACCEPTABLE \cr
HTTP\_PROXY\_AUTHENTICATION\_REQUIRED \cr
HTTP\_REQUEST\_TIME\_OUT \cr
HTTP\_CONFLICT \cr
HTTP\_GONE \cr
HTTP\_LENGTH\_REQUIRED \cr
HTTP\_PRECONDITION\_FAILED \cr
HTTP\_REQUEST\_ENTITY\_TOO\_LARGE \cr
HTTP\_REQUEST\_URI\_TOO\_LARGE \cr
HTTP\_UNSUPPORTED\_MEDIA\_TYPE \cr
HTTP\_RANGE\_NOT\_SATISFIABLE \cr
HTTP\_EXPECTATION\_FAILED \cr
HTTP\_UNPROCESSABLE\_ENTITY \cr
HTTP\_LOCKED \cr
HTTP\_FAILED\_DEPENDENCY \cr
HTTP\_UPGRADE\_REQUIRED \cr
HTTP\_INTERNAL\_SERVER\_ERROR \cr
HTTP\_NOT\_IMPLEMENTED \cr
HTTP\_BAD\_GATEWAY \cr
HTTP\_SERVICE\_UNAVAILABLE \cr
HTTP\_GATEWAY\_TIME\_OUT \cr
HTTP\_VERSION\_NOT\_SUPPORTED \cr
HTTP\_VARIANT\_ALSO\_VARIES \cr
HTTP\_INSUFFICIENT\_STORAGE \cr
HTTP\_NOT\_EXTENDED \cr
}
\alias{DONE}
\alias{DECLINED}
\alias{OK}
\alias{HTTP_CONTINUE}
\alias{HTTP_SWITCHING_PROTOCOLS}
\alias{HTTP_PROCESSING}
\alias{HTTP_OK}
\alias{HTTP_CREATED}
\alias{HTTP_ACCEPTED}
\alias{HTTP_NON_AUTHORITATIVE}
\alias{HTTP_NO_CONTENT}
\alias{HTTP_RESET_CONTENT}
\alias{HTTP_PARTIAL_CONTENT}
\alias{HTTP_MULTI_STATUS}
\alias{HTTP_MULTIPLE_CHOICES}
\alias{HTTP_MOVED_PERMANENTLY}
\alias{HTTP_MOVED_TEMPORARILY}
\alias{HTTP_SEE_OTHER}
\alias{HTTP_NOT_MODIFIED}
\alias{HTTP_USE_PROXY}
\alias{HTTP_TEMPORARY_REDIRECT}
\alias{HTTP_BAD_REQUEST}
\alias{HTTP_UNAUTHORIZED}
\alias{HTTP_PAYMENT_REQUIRED}
\alias{HTTP_FORBIDDEN}
\alias{HTTP_NOT_FOUND}
\alias{HTTP_METHOD_NOT_ALLOWED}
\alias{HTTP_NOT_ACCEPTABLE}
\alias{HTTP_PROXY_AUTHENTICATION_REQUIRED}
\alias{HTTP_REQUEST_TIME_OUT}
\alias{HTTP_CONFLICT}
\alias{HTTP_GONE}
\alias{HTTP_LENGTH_REQUIRED}
\alias{HTTP_PRECONDITION_FAILED}
\alias{HTTP_REQUEST_ENTITY_TOO_LARGE}
\alias{HTTP_REQUEST_URI_TOO_LARGE}
\alias{HTTP_UNSUPPORTED_MEDIA_TYPE}
\alias{HTTP_RANGE_NOT_SATISFIABLE}
\alias{HTTP_EXPECTATION_FAILED}
\alias{HTTP_UNPROCESSABLE_ENTITY}
\alias{HTTP_LOCKED}
\alias{HTTP_FAILED_DEPENDENCY}
\alias{HTTP_UPGRADE_REQUIRED}
\alias{HTTP_INTERNAL_SERVER_ERROR}
\alias{HTTP_NOT_IMPLEMENTED}
\alias{HTTP_BAD_GATEWAY}
\alias{HTTP_SERVICE_UNAVAILABLE}
\alias{HTTP_GATEWAY_TIME_OUT}
\alias{HTTP_VERSION_NOT_SUPPORTED}
\alias{HTTP_VARIANT_ALSO_VARIES}
\alias{HTTP_INSUFFICIENT_STORAGE}
\alias{HTTP_NOT_EXTENDED}
\keyword{programming}
